Refactor(Shared) : 멘토링 공유용 코드 작성
