+++ src/components/css/index.css
@@ -1,7 +1,3 @@
-html {
-  min-width: 375px;
-}
-
 body {
   box-sizing: border-box;
   margin: 0;
